Added utils::property<T>.
Added utils::property<T> to simulate a C# style property in C++ compiler.

Added the UTILS_ prefix before all defined macros.

Added a macro that can be used to make a function non-callable, aka. delete it in C++14 compiler, or make it private and not define it in C++11 or lower. NOTE: You must put the function in the scope defined by UTILS_DELETE_FUCTION_SCOPE if it is to be used in a class.